+++ dumps/hcl/heredoc.tf
@@ -1,0 +1,8 @@
+locals {
+  a =<<END
+    line 1
+    line 2
+    ${var.g}
+    line 3
+  END
+}